fintech_agent/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.84.0
faiss-cpu>1.8.0
numpy==2.3.3
python-dotenv==1.0.1
